Rename vscode related things
