+++ package.json
@@ -46,10 +46,12 @@
     ]
   },
   "scripts": {
-    "build": "node src/buildTheme.ts",
-    "watch": "node --watch-path=./src ./src/buildTheme.ts"
+    "build": "node src/main.ts",
+    "watch": "node --watch-path=./src ./src/main.ts"
   },
   "devDependencies": {
+    "@types/mustache": "^4.2.6",
+    "@types/node": "^22.17.0",
     "@vscode/vsce": "^3.3.0",
     "json5": "^2.2.3",
     "mustache": "^4.2.0",
