+++ src/main.ts
@@ -1,0 +1,3 @@
+import { buildVsCodeThemes } from "./vscode/buildThemes.ts";
+
+buildVsCodeThemes();
